pyproject.toml
[project]
name = "neckarwave-scripts"
version = "0.1.0"
description = "Business analysis and management scripts for Neckar Wave"
authors = [
    {name = "Andreas Schmidt", email = ""},
]
dependencies = [
    "pandas>=2.0.3",
    "openpyxl>=3.1.2",
    "seaborn>=0.13.0",
    "matplotlib>=3.10.3",
    "ace-tools>=0.0",
    "python-dotenv>=1.1.0",
    "XlsxWriter>=3.2.3",
    "streamlit>=1.46.1",
    # "openai-whisper",
    "openai",
    "colorlog>=6.10.1",
    "pydatev>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {text = "MIT"}

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.5",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["I"]
